requirements.txt
discord.py
Pillow
DateTime
topggpy